# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + MCP 协议的企业级零信任安全智能体平台 (AI SecOps)",
        "title": "ZeroTrust Sentinel API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Phantomor",
            "url": "https://github.com/Phantomor"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/user/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "清除客户端的 login_token Cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块 (User)"
                ],
                "summary": "用户退出登录",
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": \"退出成功\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/user/get/login": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "从上下文中获取当前登录用户的 ID 和账号信息（需要登录凭证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块 (User)"
                ],
                "summary": "获取当前登录用户信息",
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": {\"id\": 1, \"userAccount\": \"xxx\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "通过账号密码登录并返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "账号或密码错误",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "注册新用户，账号至少4位，密码至少6位。密码将在入库前通过 Bcrypt 加密。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块 (User)"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": \"注册成功\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"code\": 400, \"message\": \"账号已存在或参数校验失败\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"code\": 500, \"message\": \"服务端异常(如加密失败或数据库写入失败)\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "userAccount": {
                    "type": "string",
                    "example": "admin"
                },
                "userPassword": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "userAccount": {
                    "type": "string",
                    "example": "admin"
                },
                "userName": {
                    "type": "string",
                    "example": "管理员"
                },
                "userRole": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "userAccount": {
                    "type": "string",
                    "minLength": 4,
                    "example": "testuser"
                },
                "userPassword": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                }
            }
        },
...
        "pkg.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "这个字段名或 JSON tag 必须叫 data"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}